order_service/requirements.txt
fastapi[all]
SQLAlchemy
asyncpg
aio_pika
celery
